[serializer] do not copy code if snapshot is not required to be deterministic.
Code in V8 embeds memory addresses. Upon deserialization, those addresses will
be updated anyways, so it's not important whether the serializer records those
addresses. In order to have a deterministic build, we create a copy of the code
and null out memory addresses and serialize this copy.

For the code cache, we do not care about determinism. By avoiding the copy we
can save some time spent on serialization. For the mandreel script in Octane,
this reduces the serialization time from ~85ms to ~75ms.

R=vogelheim@chromium.org

Review-Url: https://codereview.chromium.org/1991033003
Cr-Commit-Position: refs/heads/master@{#36350}