[wasm] Delay throwing an error until the sequential phase of the compilation.
Signalling an error to the ErrorThrower causes a heap allocation and should
therefore not happen in the parallel phase of the compilation.

Additionally I did some refactoring: I pulled the function
BuildGraphForWasmFunction into the WasmCompilationUnit and as a consequence
I can use the class fields directly instead of passing them as parameters.

BUG=chromium::614291
LOG=N
R=titzer@chromium.org, bradnelson@chromium.org

Review-Url: https://codereview.chromium.org/2007113004
Cr-Commit-Position: refs/heads/master@{#36515}